requirements.txt
python-telegram-bot
langchain-community
requests
